+++ experimental/packages/opentelemetry-instrumentation-http/test/functionals/http-metrics.test.ts
@@ -50,7 +50,7 @@ instrumentation.disable();
 import * as http from 'http';
 import { SemconvStability } from '../../src/internal-types';
 import { getRPCMetadata, RPCType } from '@opentelemetry/core';
-import { AsyncHooksContextManager } from '@opentelemetry/context-async-hooks';
+import { AsyncLocalStorageContextManager } from '@opentelemetry/context-async-local';
 
 let server: http.Server;
 const serverPort = 22346;
@@ -71,7 +71,7 @@ describe('metrics', () => {
   let contextManager: ContextManager;
 
   beforeEach(() => {
-    contextManager = new AsyncHooksContextManager().enable();
+    contextManager = new AsyncLocalStorageContextManager().enable();
     context.setGlobalContextManager(contextManager);
     instrumentation['_updateMetricInstruments']();
     metricsMemoryExporter.reset();
