+++ experimental/packages/opentelemetry-instrumentation-http/tsconfig.json
@@ -13,9 +13,6 @@
       "path": "../../../api"
     },
     {
-      "path": "../../../packages/opentelemetry-context-async-hooks"
-    },
-    {
       "path": "../../../packages/opentelemetry-core"
     },
     {
