+++ packages/opentelemetry-sdk-trace-node/tsconfig.json
@@ -16,9 +16,6 @@
       "path": "../../semantic-conventions"
     },
     {
-      "path": "../opentelemetry-context-async-hooks"
-    },
-    {
       "path": "../opentelemetry-core"
     },
     {
